+++ pappl/base-private.h
@@ -53,10 +53,6 @@ extern char **environ;
 #  define _PAPPL_LOOKUP_STRING(bit,strings) _papplLookupString(bit, sizeof(strings) / sizeof(strings[0]), strings)
 #  define _PAPPL_LOOKUP_VALUE(keyword,strings) _papplLookupValue(keyword, sizeof(strings) / sizeof(strings[0]), strings)
 
-#  ifndef HAVE_STRLCPY
-#    define strlcpy(dst,src,dstsize) _pappl_strlcpy(dst,src,dstsize)
-#  endif // !HAVE_STRLCPY
-
 
 //
 // Types and structures...
@@ -80,14 +76,9 @@ typedef struct _pappl_link_s		// Web interface navigation link
 // Utility functions...
 //
 
-#  ifndef HAVE_STRLCPY
-extern size_t		_pappl_strlcpy(char *dst, const char *src, size_t dstsize) _PAPPL_PRIVATE;
-#  endif // !HAVE_STRLCPY
 extern ipp_t		*_papplContactExport(pappl_contact_t *contact) _PAPPL_PRIVATE;
 extern void		_papplContactImport(ipp_t *col, pappl_contact_t *contact) _PAPPL_PRIVATE;
 extern void		_papplCopyAttributes(ipp_t *to, ipp_t *from, cups_array_t *ra, ipp_tag_t group_tag, int quickcopy) _PAPPL_PRIVATE;
-extern unsigned		_papplGetRand(void) _PAPPL_PRIVATE;
-extern const char	*_papplGetTempDir(void) _PAPPL_PRIVATE;
 extern const char	*_papplLookupString(unsigned bit, size_t num_strings, const char * const *strings) _PAPPL_PRIVATE;
 extern unsigned		_papplLookupValue(const char *keyword, size_t num_strings, const char * const *strings) _PAPPL_PRIVATE;